school_project/requirements.txt
Django>=4.2.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2022.1
django-filter>=23.2
djangorestframework>=3.14.0
openpyxl>=3.1.2
Pillow>=10.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-debug-toolbar>=4.2.0
django-cors-headers>=4.3.1
